// types.ts
export type TypographyModifier =
  | 'Page'
  | 'H1'
  | 'H2'
  | 'H3'
  | 'H4'
  | 'H5'
  | 'H6'
  | 'TitleSection'
  | 'TitleBlockMedium'
  | 'TitleBlockSmall'
  | 'TitleBlockSmallUpper'
  | 'FormFieldLabel'
  | 'TextNote'
  | 'ItemPriceBefore'
  | 'ItemPriceCurrent'
  | 'ItemCommission'
  | 'ScreenTitle'
  | 'ScreenSubTitle'
  | 'Text3xLgBold'
  | 'Text2xLgBold'
  | 'TextXLgBold'
  | 'TextLgBold'
  | 'TextMdBold'
  | 'TextNmBold'
  | 'TextSmBold'
  | 'TextXsBold'
  | 'Text2XsBold'
  | 'Text3xLgMedium'
  | 'Text2xLgMedium'
  | 'TextXLgMedium'
  | 'TextLgMedium'
  | 'TextMdMedium'
  | 'TextNmMedium'
  | 'TextSmMedium'
  | 'TextXsMedium'
  | 'Text2XsMedium'
  | 'Text3xLgNormal'
  | 'Text2xLgNormal'
  | 'TextXLgNormal'
  | 'TextLgNormal'
  | 'TextMdNormal'
  | 'TextNmNormal'
  | 'TextSmNormal'
  | 'TextXsNormal'
  | 'Text2XsNormal'
  | 'Display3xLgBold'
  | 'Display2xLgBold'
  | 'DisplayXLgBold'
  | 'DisplayLgBold'
  | 'DisplayMdBold'
  | 'DisplayNmBold'
  | 'DisplaySmBold'
  | 'DisplayXsBold'
  | 'Display2XsBold'
  | 'Display3xLgMedium'
  | 'Display2xLgMedium'
  | 'DisplayXLgMedium'
  | 'DisplayLgMedium'
  | 'DisplayMdMedium'
  | 'DisplayNmMedium'
  | 'DisplaySmMedium'
  | 'DisplayXsMedium'
  | 'Display2XsMedium'
  | 'Display3xLgNormal'
  | 'Display2xLgNormal'
  | 'DisplayXLgNormal'
  | 'DisplayLgNormal'
  | 'DisplayMdNormal'
  | 'DisplayNmNormal'
  | 'DisplaySmNormal'
  | 'DisplayXsNormal'
  | 'Display2XsNormal';
// ...
export interface MarginModifiers {
  margin?: KSpacing | 0;
  marginL?: KSpacing | 0;
  marginR?: KSpacing | 0;
  marginT?: KSpacing | 0;
  marginB?: KSpacing | 0;
  marginH?: KSpacing | 0;
  marginV?: KSpacing | 0;
}
// ...
export interface TestModifier {
  testID?: string;
}
// ...
export type KRadius =
  | 'x' // 4px
  | '2x' // 8px
  | '3x' // 12px
  | '4x' // 16px
  | '6x' // 24px
  | 'round'; // round
export type KSpacing =
  | '0.25rem' // 4px
  | '0.5rem' // 8px
  | '0.75rem' // 12px
  | '1rem' // 16px
  | '1.25rem' // 20px
  | '1.5rem' // 24px
  | '2rem' // 32px
  | '2.5rem' // 40px
  | '3rem' // 48px
  | '4rem' // 64px
  | '5rem' // 80px
  | '6rem' // 96px
  | '8rem' // 128px
  | '10rem' // 160px
  | '12rem' // 192px
  | '14rem' // 234px
  | '16rem'; // 256px
export type KContainer = 'sm' | 'md' | 'lg' | 'xl';
export type KThickness = 'thin' | 'thick';
export type KButtonSize = 'xlg' | 'lg' | 'md' | 'nm' | 'sm' | 'xs';
export type KButtonKind = 'primary' | 'secondary' | 'success' | 'warning' | 'danger' | 'light';
export type KButtonLabelWeight = 'normal' | 'medium' | 'bold';
export interface KButtonProps extends TestModifier, MarginModifiers {
  br?: KRadius;
  brC?: string;
  kind?: KButtonKind;
  size?: KButtonSize;
  weight?: KButtonLabelWeight;
  label?: string;
  icon?: KButtonIconProps;
  iconAlignment?: 'left' | 'right';
  disabled?: boolean;
  loading?: boolean;
  stretch?: boolean | 'left' | 'right';
  revert?: boolean;
  thickness?: KThickness;

  onPress?: () => void;
  onLongPress?: () => void;
  enhanceStyle?: StyleProp<ViewStyle>;
  brW?: number;
}

export interface KButtonBaseProps extends KButtonProps {
  background: string;
  tintColor: string;
  isLink: boolean;
}
// ...
export type KVectorIconsProvider =
  | 'Droppii'
  | 'DroppiiNew'
  | 'Entypo'
  | 'EvilsIcons'
  | 'Feather'
  | 'FontAwesome'
  | 'Foundation'
  | 'Ionicons'
  | 'MaterialIcons'
  | 'MaterialCommunityIcons'
  | 'Octicons'
  | 'Zocial'
  | 'SimpleLineIcons';
// ...
export interface KButtonIconProps {
  vectorName?: string; // for vector icon
  vectorProvider?: KVectorIconsProvider; // for vector icon
  imageSource?: string | number; // for image icon,
  tintColor?: string;
  size?: number;
}